Major update to ActionPanel
moved panel to the side and it now displays more pertinent information

+++ src/csci_1082/finalproject/blackjack/Dealer.java
@@ -54,7 +54,7 @@ public class Dealer extends Player{
 	 */
 	public void payPlayer(Player currentPlayer, PlayerHands hand) {
 		if (hand.isBlackJack()) {
-			currentPlayer.addMoney(hand.getHandBet() * GameEngine.BLACKJACK_PAYOUT);
+			currentPlayer.addMoney((int)(hand.getHandBet() * GameEngine.BLACKJACK_PAYOUT));
 		} else {
 		currentPlayer.addMoney(hand.getHandBet() * 2);
 		}
